+++ app/assets/javascripts/app/views/single-post-viewer/single_post_interaction_counts.js
@@ -4,8 +4,15 @@ app.views.SinglePostInteractionCounts = app.views.Base.extend({
   templateName: "single-post-viewer/single-post-interaction-counts",
   tooltipSelector: ".avatar.micro",
 
+  events: {
+    "click #show-all-likes": "showAllLikes",
+    "click #show-all-reshares": "showAllReshares"
+  },
+
   initialize: function() {
     this.model.interactions.on("change", this.render, this);
+    this.model.interactions.likes.on("change", this.render, this);
+    this.model.interactions.reshares.on("change", this.render, this);
   },
 
   presenter: function() {
@@ -15,8 +22,28 @@ app.views.SinglePostInteractionCounts = app.views.Base.extend({
       reshares: interactions.reshares.toJSON(),
       commentsCount: interactions.commentsCount(),
       likesCount: interactions.likesCount(),
-      resharesCount: interactions.resharesCount()
+      resharesCount: interactions.resharesCount(),
+      showMoreLikes: interactions.likes.length < interactions.likesCount(),
+      showMoreReshares: interactions.reshares.length < interactions.resharesCount()
     };
+  },
+
+  _showAll: function(interactionType, models) {
+    this.$("#show-all-" + interactionType).addClass("hidden");
+    this.$("#" + interactionType + " .loader").removeClass("hidden");
+    models.fetch({success: function() {
+      models.trigger("change");
+    }});
+  },
+
+  showAllLikes: function(evt) {
+    evt.preventDefault();
+    this._showAll("likes", this.model.interactions.likes);
+  },
+
+  showAllReshares: function(evt) {
+    evt.preventDefault();
+    this._showAll("reshares", this.model.interactions.reshares);
   }
 });
 // @license-end